move enable tma to env_util
