+++ tzrec/ops/utils.py
@@ -10,31 +10,9 @@
 # limitations under the License.
 
 import copy
-import os
 from typing import List
 
 import torch
-
-from tzrec.utils.logging_util import logger
-
-
-def enable_tma() -> bool:
-    flag = os.environ.get("ENABLE_TMA", "0") == "1"
-    if flag:
-        if torch.cuda.is_available():
-            if torch.cuda.get_device_capability(torch.device("cuda"))[0] >= 9:
-                import triton
-                from packaging import version
-
-                if version.parse(triton.__version__) >= version.parse("3.5.0"):
-                    return True
-                else:
-                    logger.warning("triton version lower than 3.5.0, we disable TMA.")
-            else:
-                logger.warning("device capability lower than 9.0, we disable TMA.")
-        else:
-            logger.warning("CUDA not available, we disable TMA.")
-    return False
 
 
 def switch_to_contiguous_if_needed(x: torch.Tensor) -> torch.Tensor:
